agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tellaro-agent"
version = "0.1.0"
description = "Tellaro PM Agent - manages Claude Code instances on behalf of a user"
requires-python = ">=3.12"
dependencies = [
    "websockets>=14,<15",
    "pydantic>=2.10,<3",
    "pydantic-settings>=2.7,<3",
    "python-dotenv>=1.0,<2",
    "httpx>=0.28,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.25,<1",
    "pytest-cov>=6,<7",
    "ruff>=0.9,<1",
    "pyright>=1.1,<2",
]

[project.scripts]
tellaro-agent = "tellaro_agent.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/tellaro_agent"]

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "C4", "SIM", "TCH", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["tellaro_agent"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
venvPath = "."
venv = ".venv"
include = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
